# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API modular em Go com Gin e Swagger.",
        "title": "Backend Go API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Suporte API",
            "url": "http://www.swagger.io/support",
            "email": "suporte@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8888",
    "basePath": "/v1",
    "paths": {
...
        "/role": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Retorna papéis com paginação e filtros.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Role"
                ],
                "summary": "Listar Papéis (Roles)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Número da página (padrão: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Itens por página (padrão: 25)",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Termo de busca",
                        "name": "searchWord",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Campos para busca (ex: name)",
                        "name": "searchFields",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Campo para ordenação",
                        "name": "orderBy",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Direção da ordenação (asc/desc)",
                        "name": "orderDirection",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_app_role.RoleListResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Cria um novo papel com permissões específicas.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Role"
                ],
                "summary": "Criar Papel (Role)",
                "parameters": [
                    {
                        "description": "Dados do papel",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_app_role.CreateRoleDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/backend-go_internal_core_models.Role"
                        }
                    },
                    "400": {
                        "description": "Dados inválidos",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/role/all": {
            "get": {
                "description": "Retorna todos os papéis (usado para selects/lookups).",
                "tags": [
                    "Role"
                ],
                "summary": "Listar Todos os Papéis (Roles)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Termo de busca",
                        "name": "searchWord",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Campos para busca",
                        "name": "searchFields",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Campo para ordenação",
                        "name": "orderBy",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Direção da ordenação",
                        "name": "orderDirection",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_app_role.RoleListAllResponse"
                        }
                    }
                }
            }
        },
        "/role/features": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Retorna a lista de todas as funcionalidades disponíveis para atribuição de permissões.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Role"
                ],
                "summary": "Listar Funcionalidades",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/backend-go_internal_core_models.Feature"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/role/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Retorna os detalhes de um papel e suas permissões.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Role"
                ],
                "summary": "Obter Papel (Role)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do Papel",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend-go_internal_core_models.Role"
                        }
                    },
                    "404": {
                        "description": "Papel não encontrado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Atualiza os dados e permissões de um papel baseado no ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Role"
                ],
                "summary": "Atualizar Papel (Role)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do Papel",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados do papel",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_app_role.CreateRoleDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend-go_internal_core_models.Role"
                        }
                    },
                    "400": {
                        "description": "Dados inválidos",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Remove um papel baseado no ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Role"
                ],
                "summary": "Excluir Papel (Role)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do Papel",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message: papel excluído com sucesso",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/role/{id}/status": {
            "patch": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Define se um papel está ativo ou inativo.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Role"
                ],
                "summary": "Alterar Status do Papel (Role)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do Papel",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Novo status",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_app_role.UpdateStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message: status atualizado com sucesso",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "backend-go_internal_core_models.Feature": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_deleted": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "backend-go_internal_core_models.Role": {
            "type": "object",
            "properties": {
                "RoleFeature": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/backend-go_internal_core_models.RoleFeature"
                    }
                },
                "active": {
                    "type": "boolean"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_deleted": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "backend-go_internal_core_models.RoleFeature": {
            "type": "object",
            "properties": {
                "activate": {
                    "type": "boolean"
                },
                "create": {
                    "type": "boolean"
                },
                "delete": {
                    "type": "boolean"
                },
                "id_feature": {
                    "type": "string"
                },
                "id_role": {
                    "type": "string"
                },
                "view": {
                    "type": "boolean"
                }
            }
        },
...
        "internal_app_role.CreateRoleDTO": {
            "type": "object",
            "required": [
                "description",
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "permissions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/backend-go_internal_core_models.RoleFeature"
                    }
                }
            }
        },
        "internal_app_role.RoleListAllResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/backend-go_internal_core_models.Role"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "internal_app_role.RoleListResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/backend-go_internal_core_models.Role"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "internal_app_role.UpdateStatusRequest": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}